Refactored all POJOs into the pojo directory. Javadocs in progress.
